* Made some progress on the Telnet listener. Now special telnet chars ( > 0x7F) are handled properly and won't be in the ASCII string. Just a few more tweaks to the way some ASCII chars are handled and it will be ready...  * Fixed NullPointerException in SkillDataLoader.  * Fixed MCDB loading of mobs and skills.  * Added classes for inventory (items, equips, pets, rings, etc.) and moved Player class from argonms.game to argonms.character.  * Support for KvjCompiler r14. Changes include Character.wz compiling and reduced size of Mob files.  * Character.wz is now processed in KvjItemDataLoader. Taming Mob IDs are now read in McdbItemDataLoader as well. In McdbItemDataLoader, equips can now be made tradeBlocked.

+++ trunk/src/argonms/loading/mob/KvjMobDataLoader.java
@@ -200,8 +200,8 @@ public class KvjMobDataLoader extends MobDataLoader {
 
 	private void processSkill(LittleEndianReader reader, MobStats stats) {
 		Skill s = new Skill();
-		s.setSkill(reader.readInt());
-		s.setLevel(reader.readInt());
+		s.setSkill(reader.readShort());
+		s.setLevel(reader.readByte());
 		stats.addSkill(s);
 	}
 }
